chore(zeebe-provider): rename to Input- & OutputProps
Let's follow the structure.
