+++ test/spec/bpmn-properties-panel/provider/zeebe/InputProps.bpmn
