+++ test/spec/bpmn-properties-panel/provider/zeebe/OutputProps.bpmn
